requirements.txt
Django==2.0.3
whitenoise==3.2.1
gunicorn==19.6.0
bson
psycopg2==2.7.5
google-cloud-storage
django-rest-framework==0.1.0
dj-database-url
django-postgresql
django-cors-headers
